# docs/swagger.yaml
basePath: /api
definitions:
...
  domain.ChargingTariff:
    properties:
      description:
        type: string
      id:
        type: integer
      image_url:
        type: string
      is_deleted:
        type: boolean
      nameof_tariff:
        type: string
      power:
        description: мощность зарядки в кВт
        type: number
      price_per_hour:
        type: number
    type: object
...
  domain.TariffRequest:
    description: Charging tariff data
    properties:
      description:
        type: string
      nameof_tariff:
        type: string
      power:
        type: number
      price_per_hour:
        type: number
    required:
    - description
    - nameof_tariff
    - power
    - price_per_hour
    type: object
  domain.User:
    properties:
      id:
        type: integer
      is_moderator:
        type: boolean
      login:
        type: string
      password:
        type: string
    type: object
  domain.UserRequest:
    description: User login and password for authentication
    properties:
      login:
        type: string
      password:
        type: string
    required:
    - login
    - password
    type: object
host: localhost:8080
info:
  contact: {}
  description: API for electric vehicle charging station management
  title: Charging Station API
  version: "1.0"
paths:
...
  /tariffs:
    get:
      consumes:
      - application/json
      description: Get list of charging tariffs with optional filtering
      parameters:
      - description: Filter by tariff name
        in: query
        name: tariffName
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.ChargingTariff'
            type: array
        "500":
          description: Internal server error
          schema:
            type: object
      summary: Get tariffs
      tags:
      - tariffs
    post:
      consumes:
      - application/json
      description: Create new charging tariff (moderator only)
      parameters:
      - description: Tariff data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.TariffRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created tariff ID
          schema:
            type: integer
        "400":
          description: Bad request
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "403":
          description: Forbidden
          schema:
            type: object
        "500":
          description: Internal server error
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Create tariff
      tags:
      - tariffs
  /tariffs/{id}:
    delete:
      consumes:
      - application/json
      description: Delete charging tariff (soft delete) (moderator only)
      parameters:
      - description: Tariff ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Tariff deleted successfully
          schema:
            type: object
        "400":
          description: Bad request
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "403":
          description: Forbidden
          schema:
            type: object
        "404":
          description: Tariff not found
          schema:
            type: object
        "500":
          description: Internal server error
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Delete tariff
      tags:
      - tariffs
    get:
      consumes:
      - application/json
      description: Get specific charging tariff by ID
      parameters:
      - description: Tariff ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.ChargingTariff'
        "400":
          description: Bad request
          schema:
            type: object
        "404":
          description: Tariff not found
          schema:
            type: object
        "500":
          description: Internal server error
          schema:
            type: object
      summary: Get tariff by ID
      tags:
      - tariffs
    put:
      consumes:
      - application/json
      description: Update existing charging tariff (moderator only)
      parameters:
      - description: Tariff ID
        in: path
        name: id
        required: true
        type: integer
      - description: Tariff data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.TariffRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Tariff updated successfully
          schema:
            type: object
        "400":
          description: Bad request
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "403":
          description: Forbidden
          schema:
            type: object
        "404":
          description: Tariff not found
          schema:
            type: object
        "500":
          description: Internal server error
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Update tariff
      tags:
      - tariffs
  /tariffs/{id}/image:
    post:
      consumes:
      - multipart/form-data
      description: Upload image for charging tariff (moderator only)
      parameters:
      - description: Tariff ID
        in: path
        name: id
        required: true
        type: integer
      - description: Tariff image
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Image uploaded successfully
          schema:
            type: object
        "400":
          description: Bad request
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "403":
          description: Forbidden
          schema:
            type: object
        "404":
          description: Tariff not found
          schema:
            type: object
        "500":
          description: Internal server error
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Upload tariff image
      tags:
      - tariffs
  /users/{id}:
    get:
      consumes:
      - application/json
      description: Get user profile by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.User'
        "400":
          description: Bad request
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "404":
          description: User not found
          schema:
            type: object
        "500":
          description: Internal server error
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Get user by ID
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update user profile
      parameters:
      - description: User id
        in: path
        name: id
        required: true
        type: integer
      - description: User data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.UserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: User updated successfully
          schema:
            type: object
        "400":
          description: Bad request
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "404":
          description: User not found
          schema:
            type: object
        "409":
          description: Conflict - already exists
          schema:
            type: object
        "500":
          description: Internal server error
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Update user
      tags:
      - users
  /users/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and get JWT token
      parameters:
      - description: User credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.UserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login response with token
          schema:
            type: object
        "400":
          description: Bad request
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "500":
          description: Internal server error
          schema:
            type: object
      summary: Login user
      tags:
      - users
  /users/logout:
    post:
      consumes:
      - application/json
      description: Invalidate user's JWT token
      produces:
      - application/json
      responses:
        "200":
          description: Logout successful
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "500":
          description: Internal server error
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Logout user
      tags:
      - users
  /users/register:
    post:
      consumes:
      - application/json
      description: Register new user account
      parameters:
      - description: User credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.UserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created id
          schema:
            type: integer
        "400":
          description: Bad request
          schema:
            type: object
        "409":
          description: Conflict - already exists
          schema:
            type: object
        "500":
          description: Internal server error
          schema:
            type: object
      summary: Register user
      tags:
      - users
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"